auth/requirements.txt
fastapi[standard]
pyjwt
pydantic-settings
psycopg[binary,pool]
redis